requirements.txt
Django~=2.2.4
django_compressor
numpy
Pillow
django-htmlmin
